Removed jUnit, bash-based build.sh and examples/lwjgl/run.sh, JNI to load directly from FS

+++ src/main/java/org/jetbrains/skija/JNI.java
@@ -3,6 +3,8 @@ package org.jetbrains.skija;
 import java.io.File;
 import java.io.InputStream;
 import java.io.IOException;
+import java.net.URL;
+import java.net.URISyntaxException;
 import java.nio.file.Files;
 import java.nio.file.StandardCopyOption;
 
@@ -13,24 +15,31 @@ public class JNI {
     public static void loadLibrary(String resourcePath, String name) {
         if (loaded) return;
         try {
-            File tempDir = new File(System.getProperty("java.io.tmpdir"), "skija_" + System.nanoTime());
-            tempDir.mkdirs();
-            tempDir.deleteOnExit();
-
+            File file;
             String fileName = "lib" + name + "." + libraryExtension();
-            File temp = new File(tempDir, fileName);
-            temp.deleteOnExit();
 
-            try (InputStream is = JNI.class.getResourceAsStream(resourcePath + fileName)) {
-                if (is != null)
-                    Files.copy(is, temp.toPath(), StandardCopyOption.REPLACE_EXISTING);
-                else
-                    throw new IllegalArgumentException(
-                        "Library " + fileName + " is not found in " + resourcePath
-                    );
-            }
+            URL url = JNI.class.getResource(resourcePath + fileName);
+            if (url == null)
+                throw new IllegalArgumentException("Library " + fileName + " is not found in " + resourcePath);
+            else if (url.getProtocol() == "file")
+                try {
+                    // System.out.println("Loading " + url);
+                    file = new File(url.toURI());
+                } catch (URISyntaxException e){
+                    throw new RuntimeException(e);
+                }
+            else
+                try (InputStream is = url.openStream()) {
+                    File tempDir = new File(System.getProperty("java.io.tmpdir"), "skija_" + System.nanoTime());
+                    tempDir.mkdirs();
+                    tempDir.deleteOnExit();
+                    file = new File(tempDir, fileName);
+                    file.deleteOnExit();
+                    Files.copy(is, file.toPath(), StandardCopyOption.REPLACE_EXISTING);
+                    // System.out.println("Loading " + url + " from " + file);
+                }
 
-            System.load(temp.getAbsolutePath());
+            System.load(file.getAbsolutePath());
             loaded = true;
         } catch (IOException e) {
             throw new RuntimeException(e);
